Set up default save creation (in-editor only)
Also added a script to remove objects if they're supposed to be in-editor only
