+++ Assets/Resources/Saves/default.dat.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6653fe48736f607448067b6d14c4f1d6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
