--- Assets/Scripts/LoadingScreenCreature.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dd58a796a998bf64588b4d6f975a0a7f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
