--- Assets/Scripts/ScreenLoading.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0f0a3c48dd9afdd469fa68c9e95ed5eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
