+++ Assets/Scripts/UI/Loading.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f4e6101801023743a167f4daf925c5b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
