+++ Assets/Scripts/Utility/EditorOnly.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93827384990d192449f379d2f9690b9d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: -10000
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
