+++ ProjectSettings/ProjectSettings.asset
@@ -134,7 +134,7 @@ PlayerSettings:
     16:10: 1
     16:9: 1
     Others: 0
-  bundleVersion: 0.2.86
+  bundleVersion: 0.2.87
   preloadedAssets:
   - {fileID: 0}
   - {fileID: 0}
